% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.mrccc_fit}
\alias{plot.mrccc_fit}
\title{Plot the receptor-modulated ligand effect of an MR-CCC fit}
\usage{
\method{plot}{mrccc_fit}(x, Z_observed = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{mrccc_fit} returned by \code{\link[=mr_ccc]{mr_ccc()}}.}

\item{Z_observed}{Optional numeric vector of observed receptor expression
(on the original scale) used to set the plotting range. If \code{NULL}, the
range stored in the fit is used.}

\item{...}{Further arguments passed to \code{\link[=plot_effect_curve]{plot_effect_curve()}}.}
}
\value{
A \code{ggplot} object.
}
\description{
Method for \code{\link[=plot]{plot()}} that dispatches to \code{\link[=plot_effect_curve]{plot_effect_curve()}}.
}
\examples{
sim <- simulate_mrccc(n = 150, seed = 1)
fit <- mr_ccc(sim$X, sim$Z, sim$Y, sim$G, sim$H, sim$V,
              n_iter = 600, burn_in = 100, seed = 1)
plot(fit)

}
\seealso{
\code{\link[=plot_effect_curve]{plot_effect_curve()}}
}
